test: move group use case tests
