+++ backend/tests/usecases/group/AddGroupResponsibleUseCase.test.ts
@@ -1,5 +1,5 @@
 import { mockDeep, DeepMockProxy } from 'jest-mock-extended';
-import { RemoveGroupResponsibleUseCase } from '../../../usecases/userGroup/RemoveGroupResponsibleUseCase';
+import { AddGroupResponsibleUseCase } from '../../../usecases/group/AddGroupResponsibleUseCase';
 import { UserGroupRepositoryPort } from '../../../domain/ports/UserGroupRepositoryPort';
 import { UserRepositoryPort } from '../../../domain/ports/UserRepositoryPort';
 import { UserGroup } from '../../../domain/entities/UserGroup';
@@ -11,10 +11,10 @@ import { Permission } from '../../../domain/entities/Permission';
 import { PermissionKeys } from '../../../domain/entities/PermissionKeys';
 import { PermissionChecker } from '../../../domain/services/PermissionChecker';
 
-describe('RemoveGroupResponsibleUseCase', () => {
+describe('AddGroupResponsibleUseCase', () => {
   let groupRepo: DeepMockProxy<UserGroupRepositoryPort>;
   let userRepo: DeepMockProxy<UserRepositoryPort>;
-  let useCase: RemoveGroupResponsibleUseCase;
+  let useCase: AddGroupResponsibleUseCase;
   let checker: PermissionChecker;
   let site: Site;
   let dept: Department;
@@ -41,20 +41,21 @@ describe('RemoveGroupResponsibleUseCase', () => {
       [new Permission('p', PermissionKeys.MANAGE_GROUP_RESPONSIBLES, '')],
     );
     checker = new PermissionChecker(actor);
-    useCase = new RemoveGroupResponsibleUseCase(groupRepo, userRepo, checker);
+    useCase = new AddGroupResponsibleUseCase(groupRepo, userRepo, checker);
     user = new User('u', 'John', 'Doe', 'john@example.com', [role], 'active', dept, site);
     group = new UserGroup('g', 'Group', [user], [user]);
   });
 
-  it('should remove responsible from group', async () => {
+  it('should add responsible to group', async () => {
+    const other = new User('u2', 'Jane', 'Doe', 'jane@example.com', [role], 'active', dept, site);
     groupRepo.findById.mockResolvedValue(group);
-    userRepo.findById.mockResolvedValue(user);
-    groupRepo.removeResponsible.mockResolvedValue(group);
+    userRepo.findById.mockResolvedValue(other);
+    groupRepo.addResponsible.mockResolvedValue(group);
 
-    const result = await useCase.execute('g', 'u');
+    const result = await useCase.execute('g', 'u2');
 
     expect(result).toBe(group);
-    expect(groupRepo.removeResponsible).toHaveBeenCalledWith('g', 'u');
+    expect(groupRepo.addResponsible).toHaveBeenCalledWith('g', 'u2');
   });
 
   it('should return null when group or user missing', async () => {
@@ -64,6 +65,6 @@ describe('RemoveGroupResponsibleUseCase', () => {
     const result = await useCase.execute('g', 'u');
 
     expect(result).toBeNull();
-    expect(groupRepo.removeResponsible).not.toHaveBeenCalled();
+    expect(groupRepo.addResponsible).not.toHaveBeenCalled();
   });
 });
