+++ backend/tests/usecases/group/AddGroupUserUseCase.test.ts
@@ -1,5 +1,5 @@
 import { mockDeep, DeepMockProxy } from 'jest-mock-extended';
-import { AddGroupResponsibleUseCase } from '../../../usecases/userGroup/AddGroupResponsibleUseCase';
+import { AddGroupUserUseCase } from '../../../usecases/group/AddGroupUserUseCase';
 import { UserGroupRepositoryPort } from '../../../domain/ports/UserGroupRepositoryPort';
 import { UserRepositoryPort } from '../../../domain/ports/UserRepositoryPort';
 import { UserGroup } from '../../../domain/entities/UserGroup';
@@ -11,10 +11,10 @@ import { Permission } from '../../../domain/entities/Permission';
 import { PermissionKeys } from '../../../domain/entities/PermissionKeys';
 import { PermissionChecker } from '../../../domain/services/PermissionChecker';
 
-describe('AddGroupResponsibleUseCase', () => {
+describe('AddGroupUserUseCase', () => {
   let groupRepo: DeepMockProxy<UserGroupRepositoryPort>;
   let userRepo: DeepMockProxy<UserRepositoryPort>;
-  let useCase: AddGroupResponsibleUseCase;
+  let useCase: AddGroupUserUseCase;
   let checker: PermissionChecker;
   let site: Site;
   let dept: Department;
@@ -38,33 +38,31 @@ describe('AddGroupResponsibleUseCase', () => {
       dept,
       site,
       undefined,
-      [new Permission('p', PermissionKeys.MANAGE_GROUP_RESPONSIBLES, '')],
+      [new Permission('p', PermissionKeys.MANAGE_GROUP_MEMBERS, '')],
     );
     checker = new PermissionChecker(actor);
-    useCase = new AddGroupResponsibleUseCase(groupRepo, userRepo, checker);
+    useCase = new AddGroupUserUseCase(groupRepo, userRepo, checker);
     user = new User('u', 'John', 'Doe', 'john@example.com', [role], 'active', dept, site);
     group = new UserGroup('g', 'Group', [user], [user]);
   });
 
-  it('should add responsible to group', async () => {
+  it('should add user to group', async () => {
     const other = new User('u2', 'Jane', 'Doe', 'jane@example.com', [role], 'active', dept, site);
     groupRepo.findById.mockResolvedValue(group);
     userRepo.findById.mockResolvedValue(other);
-    groupRepo.addResponsible.mockResolvedValue(group);
-
+    groupRepo.addUser.mockResolvedValue(group);
     const result = await useCase.execute('g', 'u2');
-
     expect(result).toBe(group);
-    expect(groupRepo.addResponsible).toHaveBeenCalledWith('g', 'u2');
+    expect(groupRepo.addUser).toHaveBeenCalledWith('g', 'u2');
   });
 
-  it('should return null when group or user missing', async () => {
+  it('should return null when group or user is missing', async () => {
     groupRepo.findById.mockResolvedValue(null);
     userRepo.findById.mockResolvedValue(user);
 
     const result = await useCase.execute('g', 'u');
 
     expect(result).toBeNull();
-    expect(groupRepo.addResponsible).not.toHaveBeenCalled();
+    expect(groupRepo.addUser).not.toHaveBeenCalled();
   });
 });
