+++ backend/tests/usecases/group/CreateUserGroupUseCase.test.ts
@@ -1,5 +1,5 @@
 import { mockDeep, DeepMockProxy } from 'jest-mock-extended';
-import { UpdateUserGroupUseCase } from '../../../usecases/userGroup/UpdateUserGroupUseCase';
+import { CreateUserGroupUseCase } from '../../../usecases/group/CreateUserGroupUseCase';
 import { UserGroupRepositoryPort } from '../../../domain/ports/UserGroupRepositoryPort';
 import { UserGroup } from '../../../domain/entities/UserGroup';
 import { User } from '../../../domain/entities/User';
@@ -10,9 +10,9 @@ import { Permission } from '../../../domain/entities/Permission';
 import { PermissionChecker } from '../../../domain/services/PermissionChecker';
 import { PermissionKeys } from '../../../domain/entities/PermissionKeys';
 
-describe('UpdateUserGroupUseCase', () => {
+describe('CreateUserGroupUseCase', () => {
   let repo: DeepMockProxy<UserGroupRepositoryPort>;
-  let useCase: UpdateUserGroupUseCase;
+  let useCase: CreateUserGroupUseCase;
   let group: UserGroup;
   let user: User;
   let permissionChecker: PermissionChecker;
@@ -21,19 +21,17 @@ describe('UpdateUserGroupUseCase', () => {
     repo = mockDeep<UserGroupRepositoryPort>();
     const site = new Site('s', 'Site');
     const dept = new Department('d', 'Dept', null, null, site);
-    const role = new Role('r', 'Role', [new Permission('p', PermissionKeys.UPDATE_GROUP, 'update')]);
+    const role = new Role('r', 'Role', [new Permission('p', PermissionKeys.CREATE_GROUP, 'create')]);
     user = new User('u', 'John', 'Doe', 'john@example.com', [role], 'active', dept, site);
     permissionChecker = new PermissionChecker(user);
-    useCase = new UpdateUserGroupUseCase(repo, permissionChecker);
+    useCase = new CreateUserGroupUseCase(repo, permissionChecker);
     group = new UserGroup('g', 'Group', [user], [user]);
   });
 
-  it('should update group via repository', async () => {
-    repo.update.mockResolvedValue(group);
-
+  it('should create group via repository', async () => {
+    repo.create.mockResolvedValue(group);
     const result = await useCase.execute(group);
-
     expect(result).toBe(group);
-    expect(repo.update).toHaveBeenCalledWith(group);
+    expect(repo.create).toHaveBeenCalledWith(group);
   });
 });
