+++ backend/tests/usecases/group/RemoveGroupUserUseCase.test.ts
@@ -1,5 +1,5 @@
 import { mockDeep, DeepMockProxy } from 'jest-mock-extended';
-import { AddGroupUserUseCase } from '../../../usecases/userGroup/AddGroupUserUseCase';
+import { RemoveGroupUserUseCase } from '../../../usecases/group/RemoveGroupUserUseCase';
 import { UserGroupRepositoryPort } from '../../../domain/ports/UserGroupRepositoryPort';
 import { UserRepositoryPort } from '../../../domain/ports/UserRepositoryPort';
 import { UserGroup } from '../../../domain/entities/UserGroup';
@@ -11,10 +11,10 @@ import { Permission } from '../../../domain/entities/Permission';
 import { PermissionKeys } from '../../../domain/entities/PermissionKeys';
 import { PermissionChecker } from '../../../domain/services/PermissionChecker';
 
-describe('AddGroupUserUseCase', () => {
+describe('RemoveGroupUserUseCase', () => {
   let groupRepo: DeepMockProxy<UserGroupRepositoryPort>;
   let userRepo: DeepMockProxy<UserRepositoryPort>;
-  let useCase: AddGroupUserUseCase;
+  let useCase: RemoveGroupUserUseCase;
   let checker: PermissionChecker;
   let site: Site;
   let dept: Department;
@@ -41,28 +41,29 @@ describe('AddGroupUserUseCase', () => {
       [new Permission('p', PermissionKeys.MANAGE_GROUP_MEMBERS, '')],
     );
     checker = new PermissionChecker(actor);
-    useCase = new AddGroupUserUseCase(groupRepo, userRepo, checker);
+    useCase = new RemoveGroupUserUseCase(groupRepo, userRepo, checker);
     user = new User('u', 'John', 'Doe', 'john@example.com', [role], 'active', dept, site);
     group = new UserGroup('g', 'Group', [user], [user]);
   });
 
-  it('should add user to group', async () => {
-    const other = new User('u2', 'Jane', 'Doe', 'jane@example.com', [role], 'active', dept, site);
+  it('should remove user from group', async () => {
     groupRepo.findById.mockResolvedValue(group);
-    userRepo.findById.mockResolvedValue(other);
-    groupRepo.addUser.mockResolvedValue(group);
-    const result = await useCase.execute('g', 'u2');
+    userRepo.findById.mockResolvedValue(user);
+    groupRepo.removeUser.mockResolvedValue(group);
+
+    const result = await useCase.execute('g', 'u');
+
     expect(result).toBe(group);
-    expect(groupRepo.addUser).toHaveBeenCalledWith('g', 'u2');
+    expect(groupRepo.removeUser).toHaveBeenCalledWith('g', 'u');
   });
 
-  it('should return null when group or user is missing', async () => {
+  it('should return null when group or user missing', async () => {
     groupRepo.findById.mockResolvedValue(null);
     userRepo.findById.mockResolvedValue(user);
 
     const result = await useCase.execute('g', 'u');
 
     expect(result).toBeNull();
-    expect(groupRepo.addUser).not.toHaveBeenCalled();
+    expect(groupRepo.removeUser).not.toHaveBeenCalled();
   });
 });
