+++ backend/tests/usecases/group/UpdateUserGroupUseCase.test.ts
@@ -1,5 +1,5 @@
 import { mockDeep, DeepMockProxy } from 'jest-mock-extended';
-import { CreateUserGroupUseCase } from '../../../usecases/userGroup/CreateUserGroupUseCase';
+import { UpdateUserGroupUseCase } from '../../../usecases/group/UpdateUserGroupUseCase';
 import { UserGroupRepositoryPort } from '../../../domain/ports/UserGroupRepositoryPort';
 import { UserGroup } from '../../../domain/entities/UserGroup';
 import { User } from '../../../domain/entities/User';
@@ -10,9 +10,9 @@ import { Permission } from '../../../domain/entities/Permission';
 import { PermissionChecker } from '../../../domain/services/PermissionChecker';
 import { PermissionKeys } from '../../../domain/entities/PermissionKeys';
 
-describe('CreateUserGroupUseCase', () => {
+describe('UpdateUserGroupUseCase', () => {
   let repo: DeepMockProxy<UserGroupRepositoryPort>;
-  let useCase: CreateUserGroupUseCase;
+  let useCase: UpdateUserGroupUseCase;
   let group: UserGroup;
   let user: User;
   let permissionChecker: PermissionChecker;
@@ -21,17 +21,19 @@ describe('CreateUserGroupUseCase', () => {
     repo = mockDeep<UserGroupRepositoryPort>();
     const site = new Site('s', 'Site');
     const dept = new Department('d', 'Dept', null, null, site);
-    const role = new Role('r', 'Role', [new Permission('p', PermissionKeys.CREATE_GROUP, 'create')]);
+    const role = new Role('r', 'Role', [new Permission('p', PermissionKeys.UPDATE_GROUP, 'update')]);
     user = new User('u', 'John', 'Doe', 'john@example.com', [role], 'active', dept, site);
     permissionChecker = new PermissionChecker(user);
-    useCase = new CreateUserGroupUseCase(repo, permissionChecker);
+    useCase = new UpdateUserGroupUseCase(repo, permissionChecker);
     group = new UserGroup('g', 'Group', [user], [user]);
   });
 
-  it('should create group via repository', async () => {
-    repo.create.mockResolvedValue(group);
+  it('should update group via repository', async () => {
+    repo.update.mockResolvedValue(group);
+
     const result = await useCase.execute(group);
+
     expect(result).toBe(group);
-    expect(repo.create).toHaveBeenCalledWith(group);
+    expect(repo.update).toHaveBeenCalledWith(group);
   });
 });
