+++ backend/usecases/group/GetGroupResponsiblesUseCase.ts
