+++ backend/usecases/group/RemoveGroupResponsibleUseCase.ts
