+++ backend/usecases/group/RemoveGroupUserUseCase.ts
