+++ backend/usecases/group/RemoveUserGroupUseCase.ts
